requirements.txt
praw
config
python-dotenv
random